pyproject.toml
[project]
name = "python-rag-docling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "datasets>=4.8.4",
    "docling>=2.81.0",
    "docling-hierarchical-pdf>=0.1.5",
    "fastapi>=0.115.0",
    "groq<1",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "semantic-router>=0.1.12",
    "sentence-transformers>=5.3.0",
    "typer>=0.21.2",
    "uvicorn>=0.32.0",
]
